refactor: Removed needless allocations and function calls, Rewrote Immuclient package layout
