+++ cmd/immuclient/commands/init.go
@@ -1,0 +1,96 @@
+/*
+Copyright 2019-2020 vChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package commands
+
+import (
+	c "github.com/codenotary/immudb/cmd"
+	"github.com/codenotary/immudb/cmd/docs/man"
+	"github.com/codenotary/immudb/pkg/client"
+	"github.com/codenotary/immudb/pkg/gw"
+	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
+)
+
+type CommandlineClient struct {
+	tokenFilename string
+	ImmuClient    client.ImmuClient
+}
+
+func (cl *CommandlineClient) Init(cmd *cobra.Command, o *c.Options) {
+	if err := configureOptions(cmd, o); err != nil {
+		c.QuitToStdErr(err)
+	}
+	// login and logout
+	cl.login(cmd)
+	cl.logout(cmd)
+	// current status
+	cl.currentRoot(cmd)
+	// get operations
+	cl.getByIndex(cmd)
+	cl.getKey(cmd)
+	cl.rawSafeGetKey(cmd)
+	cl.safeGetKey(cmd)
+	// set operations
+	cl.rawSafeSet(cmd)
+	cl.set(cmd)
+	cl.safeset(cmd)
+	cl.zAdd(cmd)
+	cl.safeZAdd(cmd)
+	// scanners
+	cl.zScan(cmd)
+	cl.iScan(cmd)
+	cl.scan(cmd)
+	cl.count(cmd)
+	// references
+	cl.reference(cmd)
+	cl.safereference(cmd)
+	// misc
+	cl.inclusion(cmd)
+	cl.consistency(cmd)
+	cl.history(cmd)
+	cl.healthCheck(cmd)
+	cl.dumpToFile(cmd)
+
+	// man file generator
+	cmd.AddCommand(man.Generate(cmd, "immuclient", "../docs/man/immuclient"))
+}
+
+func configureOptions(cmd *cobra.Command, o *c.Options) error {
+	cmd.PersistentFlags().IntP("port", "p", gw.DefaultOptions().ImmudbPort, "immudb port number")
+	cmd.PersistentFlags().StringP("address", "a", gw.DefaultOptions().ImmudbAddress, "immudb host address")
+	cmd.PersistentFlags().StringVar(&o.CfgFn, "config", "", "config file (default path are configs or $HOME. Default filename is immuclient.ini)")
+	cmd.PersistentFlags().BoolP("auth", "s", client.DefaultOptions().Auth, "use authentication")
+	if err := viper.BindPFlag("default.port", cmd.PersistentFlags().Lookup("port")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.address", cmd.PersistentFlags().Lookup("address")); err != nil {
+		return err
+	}
+	if err := viper.BindPFlag("default.auth", cmd.PersistentFlags().Lookup("auth")); err != nil {
+		return err
+	}
+	viper.SetDefault("default.port", gw.DefaultOptions().ImmudbPort)
+	viper.SetDefault("default.address", gw.DefaultOptions().ImmudbAddress)
+	viper.SetDefault("default.auth", client.DefaultOptions().Auth)
+	viper.SetDefault("default.mtls", client.DefaultOptions().MTLs)
+	viper.SetDefault("default.servername", client.DefaultMTLsOptions().Servername)
+	viper.SetDefault("default.certificate", client.DefaultMTLsOptions().Certificate)
+	viper.SetDefault("default.pkey", client.DefaultMTLsOptions().Pkey)
+	viper.SetDefault("default.clientcas", client.DefaultMTLsOptions().ClientCAs)
+
+	return nil
+}
