+++ cmd/immuclient/commands/misc.go
@@ -1,0 +1,104 @@
+/*
+Copyright 2019-2020 vChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package commands
+
+import (
+	"bytes"
+	"context"
+	"fmt"
+	"io/ioutil"
+	"os"
+	"time"
+
+	c "github.com/codenotary/immudb/cmd"
+	"github.com/spf13/cobra"
+)
+
+func (cl *CommandlineClient) history(cmd *cobra.Command) {
+	ccmd := &cobra.Command{
+		Use:     "history key",
+		Short:   "Fetch history for the item having the specified key",
+		Aliases: []string{"h"},
+		RunE: func(cmd *cobra.Command, args []string) error {
+
+			key, err := ioutil.ReadAll(bytes.NewReader([]byte(args[0])))
+			if err != nil {
+				c.QuitToStdErr(err)
+			}
+			ctx := context.Background()
+			response, err := cl.ImmuClient.History(ctx, key)
+			if err != nil {
+				c.QuitWithUserError(err)
+			}
+			for _, item := range response.Items {
+				printItem(nil, nil, item)
+				fmt.Println()
+			}
+			return nil
+		},
+		Args: cobra.ExactArgs(1),
+	}
+	cmd.AddCommand(ccmd)
+}
+
+func (cl *CommandlineClient) healthCheck(cmd *cobra.Command) {
+	ccmd := &cobra.Command{
+		Use:     "ping",
+		Short:   "Ping to check if server connection is alive",
+		Aliases: []string{"p"},
+		RunE: func(cmd *cobra.Command, args []string) error {
+
+			ctx := context.Background()
+			if err := cl.ImmuClient.HealthCheck(ctx); err != nil {
+				c.QuitWithUserError(err)
+			}
+			fmt.Println("Health check OK")
+			return nil
+		},
+		Args: cobra.NoArgs,
+	}
+	cmd.AddCommand(ccmd)
+}
+
+func (cl *CommandlineClient) dumpToFile(cmd *cobra.Command) {
+	ccmd := &cobra.Command{
+		Use:     "dump [filename]",
+		Short:   "Save a database dump to the specified filename (optional)",
+		Aliases: []string{"b"},
+		RunE: func(cmd *cobra.Command, args []string) error {
+
+			filename := fmt.Sprint("immudb_" + time.Now().Format("2006-01-02_15-04-05") + ".bkp")
+			if len(args) > 0 {
+				filename = args[0]
+			}
+			file, err := os.Create(filename)
+			defer file.Close()
+			if err != nil {
+				c.QuitToStdErr(err)
+			}
+			ctx := context.Background()
+			response, err := cl.ImmuClient.Dump(ctx, file)
+			if err != nil {
+				c.QuitWithUserError(err)
+			}
+			fmt.Printf("SUCCESS: %d key-value entries were backed-up to file %s\n", response, filename)
+			return nil
+		},
+		Args: cobra.MaximumNArgs(1),
+	}
+	cmd.AddCommand(ccmd)
+}
