+++ pkg/server/server.go
@@ -21,7 +21,6 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 	"fmt"
-	"github.com/rs/xid"
 	"io/ioutil"
 	"net"
 	"os"
@@ -29,12 +28,14 @@ import (
 	"path/filepath"
 	"syscall"
 
+	"github.com/rs/xid"
+
 	"github.com/codenotary/immudb/pkg/api/schema"
 	"github.com/codenotary/immudb/pkg/auth"
 	"github.com/codenotary/immudb/pkg/store"
 	"github.com/dgraph-io/badger/v2/pb"
 	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/grpc-ecosystem/go-grpc-middleware"
+	grpc_middleware "github.com/grpc-ecosystem/go-grpc-middleware"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/credentials"
@@ -251,6 +252,9 @@ func (s *ImmuServer) Get(ctx context.Context, k *schema.Key) (*schema.Item, erro
 
 func (s *ImmuServer) GetSV(ctx context.Context, k *schema.Key) (*schema.StructuredItem, error) {
 	it, err := s.Get(ctx, k)
+	if err != nil {
+		return nil, err
+	}
 	si, err := it.ToSItem()
 	if err != nil {
 		return nil, err
@@ -269,6 +273,9 @@ func (s *ImmuServer) SafeGet(ctx context.Context, opts *schema.SafeGetOptions) (
 
 func (s *ImmuServer) SafeGetSV(ctx context.Context, opts *schema.SafeGetOptions) (*schema.SafeStructuredItem, error) {
 	it, err := s.SafeGet(ctx, opts)
+	if err != nil {
+		return nil, err
+	}
 	ssitem, err := it.ToSafeSItem()
 	if err != nil {
 		return nil, err
@@ -293,6 +300,9 @@ func (s *ImmuServer) GetBatch(ctx context.Context, kl *schema.KeyList) (*schema.
 
 func (s *ImmuServer) GetBatchSV(ctx context.Context, kl *schema.KeyList) (*schema.StructuredItemList, error) {
 	list, err := s.GetBatch(ctx, kl)
+	if err != nil {
+		return nil, err
+	}
 	slist, err := list.ToSItemList()
 	if err != nil {
 		return nil, err
@@ -308,6 +318,9 @@ func (s *ImmuServer) Scan(ctx context.Context, opts *schema.ScanOptions) (*schem
 func (s *ImmuServer) ScanSV(ctx context.Context, opts *schema.ScanOptions) (*schema.StructuredItemList, error) {
 	s.Logger.Debugf("scan %+v", *opts)
 	list, err := s.Store.Scan(*opts)
+	if err != nil {
+		return nil, err
+	}
 	slist, err := list.ToSItemList()
 	if err != nil {
 		return nil, err
@@ -421,6 +434,9 @@ func (s *ImmuServer) ZScan(ctx context.Context, opts *schema.ZScanOptions) (*sch
 func (s *ImmuServer) ZScanSV(ctx context.Context, opts *schema.ZScanOptions) (*schema.StructuredItemList, error) {
 	s.Logger.Debugf("zscan %+v", *opts)
 	list, err := s.Store.ZScan(*opts)
+	if err != nil {
+		return nil, err
+	}
 	slist, err := list.ToSItemList()
 	if err != nil {
 		return nil, err
@@ -441,6 +457,9 @@ func (s *ImmuServer) IScan(ctx context.Context, opts *schema.IScanOptions) (*sch
 func (s *ImmuServer) IScanSV(ctx context.Context, opts *schema.IScanOptions) (*schema.SPage, error) {
 	s.Logger.Debugf("zscan %+v", *opts)
 	page, err := s.Store.IScan(*opts)
+	if err != nil {
+		return nil, err
+	}
 	SPage, err := page.ToSPage()
 	if err != nil {
 		return nil, err
